refactor(frontend): split shell messaging and tauri hotspots
