+++ openspec/changes/split-tauri-service-facade/.openspec.yaml
@@ -1,0 +1,2 @@
+schema: spec-driven
+created: 2026-04-22
