+++ src/app-shell-parts/useAppShellPromptActionsSection.ts
@@ -1,0 +1,124 @@
+// @ts-nocheck
+import { useCallback } from "react";
+import { revealItemInDir } from "@tauri-apps/plugin-opener";
+
+export function useAppShellPromptActionsSection({
+  activeWorkspace,
+  alertError,
+  connectWorkspace,
+  createPrompt,
+  deletePrompt,
+  getGlobalPromptsDir,
+  getWorkspacePromptsDir,
+  movePrompt,
+  sendUserMessageToThread,
+  startThreadForWorkspace,
+  updatePrompt,
+}: any) {
+  const handleSendPromptToNewAgent = useCallback(
+    async (text: string) => {
+      const trimmed = text.trim();
+      if (!activeWorkspace || !trimmed) {
+        return;
+      }
+      if (!activeWorkspace.connected) {
+        await connectWorkspace(activeWorkspace);
+      }
+      const threadId = await startThreadForWorkspace(activeWorkspace.id, {
+        activate: false,
+      });
+      if (!threadId) {
+        return;
+      }
+      await sendUserMessageToThread(activeWorkspace, threadId, trimmed, []);
+    },
+    [activeWorkspace, connectWorkspace, sendUserMessageToThread, startThreadForWorkspace],
+  );
+
+  const handleCreatePrompt = useCallback(
+    async (data: {
+      scope: "workspace" | "global";
+      name: string;
+      description?: string | null;
+      argumentHint?: string | null;
+      content: string;
+    }) => {
+      try {
+        await createPrompt(data);
+      } catch (error) {
+        alertError(error);
+      }
+    },
+    [alertError, createPrompt],
+  );
+
+  const handleUpdatePrompt = useCallback(
+    async (data: {
+      path: string;
+      name: string;
+      description?: string | null;
+      argumentHint?: string | null;
+      content: string;
+    }) => {
+      try {
+        await updatePrompt(data);
+      } catch (error) {
+        alertError(error);
+      }
+    },
+    [alertError, updatePrompt],
+  );
+
+  const handleDeletePrompt = useCallback(
+    async (path: string) => {
+      try {
+        await deletePrompt(path);
+      } catch (error) {
+        alertError(error);
+      }
+    },
+    [alertError, deletePrompt],
+  );
+
+  const handleMovePrompt = useCallback(
+    async (data: { path: string; scope: "workspace" | "global" }) => {
+      try {
+        await movePrompt(data);
+      } catch (error) {
+        alertError(error);
+      }
+    },
+    [alertError, movePrompt],
+  );
+
+  const handleRevealWorkspacePrompts = useCallback(async () => {
+    try {
+      const path = await getWorkspacePromptsDir();
+      await revealItemInDir(path);
+    } catch (error) {
+      alertError(error);
+    }
+  }, [alertError, getWorkspacePromptsDir]);
+
+  const handleRevealGeneralPrompts = useCallback(async () => {
+    try {
+      const path = await getGlobalPromptsDir();
+      if (!path) {
+        return;
+      }
+      await revealItemInDir(path);
+    } catch (error) {
+      alertError(error);
+    }
+  }, [alertError, getGlobalPromptsDir]);
+
+  return {
+    handleSendPromptToNewAgent,
+    handleCreatePrompt,
+    handleUpdatePrompt,
+    handleDeletePrompt,
+    handleMovePrompt,
+    handleRevealWorkspacePrompts,
+    handleRevealGeneralPrompts,
+  };
+}
