+++ src/services/tauri/dictation.ts
@@ -1,0 +1,63 @@
+import { invoke } from "@tauri-apps/api/core";
+import type {
+  DictationModelStatus,
+  DictationSessionState,
+} from "../../types";
+
+function withModelId(modelId?: string | null) {
+  return modelId ? { modelId } : {};
+}
+
+export async function getDictationModelStatus(
+  modelId?: string | null,
+): Promise<DictationModelStatus> {
+  return invoke<DictationModelStatus>(
+    "dictation_model_status",
+    withModelId(modelId),
+  );
+}
+
+export async function downloadDictationModel(
+  modelId?: string | null,
+): Promise<DictationModelStatus> {
+  return invoke<DictationModelStatus>(
+    "dictation_download_model",
+    withModelId(modelId),
+  );
+}
+
+export async function cancelDictationDownload(
+  modelId?: string | null,
+): Promise<DictationModelStatus> {
+  return invoke<DictationModelStatus>(
+    "dictation_cancel_download",
+    withModelId(modelId),
+  );
+}
+
+export async function removeDictationModel(
+  modelId?: string | null,
+): Promise<DictationModelStatus> {
+  return invoke<DictationModelStatus>(
+    "dictation_remove_model",
+    withModelId(modelId),
+  );
+}
+
+export async function startDictation(
+  preferredLanguage: string | null,
+): Promise<DictationSessionState> {
+  return invoke("dictation_start", { preferredLanguage });
+}
+
+export async function requestDictationPermission(): Promise<boolean> {
+  return invoke("dictation_request_permission");
+}
+
+export async function stopDictation(): Promise<DictationSessionState> {
+  return invoke("dictation_stop");
+}
+
+export async function cancelDictation(): Promise<DictationSessionState> {
+  return invoke("dictation_cancel");
+}
